\name{print.summary.count}
\alias{print.summary.count}
\alias{print.summary.census}
\title{Print method for summary objects}
\description{
Print method for summary objects
}
\usage{
\method{print}{summary.count}(x, ...)
}
\arguments{
  \item{x}{Object of class \dQuote{summary.count}}
  \item{\dots}{extra arguments, currently ignored}
}
\author{Robin K. S. Hankin}
\examples{
data(butterflies)
summary(butterflies)
}
\keyword{math}
