\name{saunders}
\alias{saunders}
\docType{data}
\title{Dataset due to Saunders}
\description{
A list of 20 holdfasts and a species inventory for each
}
\usage{data(saunders)}
\format{
  A data frame with 40 observations on the following 177 variables.
  \describe{
    \item{\code{exposed}}{a logical vector indicating whether the
      holdfast was exposed (\code{TRUE}) or sheltered (\code{FALSE})}
    \item{\code{Ophiactis.resiliens}}{a numeric vector giving individual counts for this species}
    \item{\code{Ophiopteris.antipodum}}{a numeric vector giving individual counts for this species}
    \item{\code{Asterina.regularis}}{a numeric vector giving individual counts for this species}
    \item{\code{Evechinus.chloroticus}}{a numeric vector giving individual counts for this species}
    \item{\code{Stichopus.juvenile.}}{a numeric vector giving individual counts for this species}
    \item{\code{Ocnus.calcarea}}{a numeric vector giving individual counts for this species}
    \item{\code{Onithochiton}}{a numeric vector giving individual counts for this species}
    \item{\code{Acanthochitona.zelandica}}{a numeric vector giving individual counts for this species}
    \item{\code{Aclis.sp..or.Seila.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Anabathron.angulatus}}{a numeric vector giving individual counts for this species}
    \item{\code{Astraea.heliotropium}}{a numeric vector giving individual counts for this species}
    \item{\code{Austromitra.rubiginosa}}{a numeric vector giving individual counts for this species}
    \item{\code{Melanella.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Buccinulum.lineum.lineum}}{a numeric vector giving individual counts for this species}
    \item{\code{Buccinulum.robustum}}{a numeric vector giving individual counts for this species}
    \item{\code{Buccinulum.sp}}{a numeric vector giving individual counts for this species}
    \item{\code{Cantharidus.purpureus}}{a numeric vector giving individual counts for this species}
    \item{\code{Crassitoniella.carinata}}{a numeric vector giving individual counts for this species}
    \item{\code{Maoricrypta.monoxyla}}{a numeric vector giving individual counts for this species}
    \item{\code{Diloma.bicanaliculata}}{a numeric vector giving individual counts for this species}
    \item{\code{Eatoniella.albocolumella}}{a numeric vector giving individual counts for this species}
    \item{\code{Eatoniella.mortoni}}{a numeric vector giving individual counts for this species}
    \item{\code{Eatoniella.olivacea}}{a numeric vector giving individual counts for this species}
    \item{\code{Eatoniella.roseocincta}}{a numeric vector giving individual counts for this species}
    \item{\code{Eatoniella.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Fictonoba.carnosa.carnosa}}{a numeric vector giving individual counts for this species}
    \item{\code{Lamellaria.ophione}}{a numeric vector giving individual counts for this species}
    \item{\code{Macrozafra.subabnormis}}{a numeric vector giving individual counts for this species}
    \item{\code{Merelina.taupoensis}}{a numeric vector giving individual counts for this species}
    \item{\code{Micrelenchus.sanguineus}}{a numeric vector giving individual counts for this species}
    \item{\code{Muricopsis.octagonus}}{a numeric vector giving individual counts for this species}
    \item{\code{Paxula.paxillus}}{a numeric vector giving individual counts for this species}
    \item{\code{Pisinna.manawatawhia}}{a numeric vector giving individual counts for this species}
    \item{\code{Pisinna.olivacea.impressa}}{a numeric vector giving individual counts for this species}
    \item{\code{Pisinna.rekohuana}}{a numeric vector giving individual counts for this species}
    \item{\code{Prolixodens.infracolor}}{a numeric vector giving individual counts for this species}
    \item{\code{Rissoina..chathamensis}}{a numeric vector giving individual counts for this species}
    \item{\code{Sagenotriphora.ampulla}}{a numeric vector giving individual counts for this species}
    \item{\code{Seila.cincta}}{a numeric vector giving individual counts for this species}
    \item{\code{Sigapatella.novaezelandiae}}{a numeric vector giving individual counts for this species}
    \item{\code{Taron.dubius}}{a numeric vector giving individual counts for this species}
    \item{\code{Trochus.viridis}}{a numeric vector giving individual counts for this species}
    \item{\code{Tugali.suteri}}{a numeric vector giving individual counts for this species}
    \item{\code{Turbo.smaragdas}}{a numeric vector giving individual counts for this species}
    \item{\code{Xymene.traversi}}{a numeric vector giving individual counts for this species}
    \item{\code{Zalipais.lissa}}{a numeric vector giving individual counts for this species}
    \item{\code{Zemitrella.choava}}{a numeric vector giving individual counts for this species}
    \item{\code{Zemitrella.pseudomarginata}}{a numeric vector giving individual counts for this species}
    \item{\code{Chemnitzia.sp}}{a numeric vector giving individual counts for this species}
    \item{\code{Linopyrga.rugata}}{a numeric vector giving individual counts for this species}
    \item{\code{Aplysia..Varria..dactylomela}}{a numeric vector giving individual counts for this species}
    \item{\code{Aphelodoris.luctuosa}}{a numeric vector giving individual counts for this species}
    \item{\code{Dendrodoris.citrina}}{a numeric vector giving individual counts for this species}
    \item{\code{Tritonia.flemingi}}{a numeric vector giving individual counts for this species}
    \item{\code{Okenia.plana}}{a numeric vector giving individual counts for this species}
    \item{\code{Borniola.reniformis}}{a numeric vector giving individual counts for this species}
    \item{\code{Cardita.brookesi}}{a numeric vector giving individual counts for this species}
    \item{\code{Chlamys.zelandiae}}{a numeric vector giving individual counts for this species}
    \item{\code{Felaniella..Zemysia..zelandica}}{a numeric vector giving individual counts for this species}
    \item{\code{Hiatella.arctica}}{a numeric vector giving individual counts for this species}
    \item{\code{Nuculidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Trichomusculus.barbatus}}{a numeric vector giving individual counts for this species}
    \item{\code{Ostracoda}}{a numeric vector giving individual counts for this species}
    \item{\code{Harpacticoid.copepods}}{a numeric vector giving individual counts for this species}
    \item{\code{Balanus.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Cumaceans}}{a numeric vector giving individual counts for this species}
    \item{\code{Mysids}}{a numeric vector giving individual counts for this species}
    \item{\code{Tanaid.1...Leptocheila}}{a numeric vector giving individual counts for this species}
    \item{\code{Tanaid.2}}{a numeric vector giving individual counts for this species}
    \item{\code{Tanaid.3...Synapseudes}}{a numeric vector giving individual counts for this species}
    \item{\code{Anatanais}}{a numeric vector giving individual counts for this species}
    \item{\code{Gnathia.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Colanthura}}{a numeric vector giving individual counts for this species}
    \item{\code{Paranthura}}{a numeric vector giving individual counts for this species}
    \item{\code{Cyathura}}{a numeric vector giving individual counts for this species}
    \item{\code{Cirolanidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Limnoriidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Plakarthrium.typicum}}{a numeric vector giving individual counts for this species}
    \item{\code{Cymodocella.capra}}{a numeric vector giving individual counts for this species}
    \item{\code{Dynamenoides.decima}}{a numeric vector giving individual counts for this species}
    \item{\code{Dynamenopsis.varicolor}}{a numeric vector giving individual counts for this species}
    \item{\code{Cassidinopsis.admirabilis}}{a numeric vector giving individual counts for this species}
    \item{\code{Asellota.sp.2}}{a numeric vector giving individual counts for this species}
    \item{\code{Janiridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Joeropsididae}}{a numeric vector giving individual counts for this species}
    \item{\code{Munnidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Neastacilla.falclandica}}{a numeric vector giving individual counts for this species}
    \item{\code{Caprella.equilibra}}{a numeric vector giving individual counts for this species}
    \item{\code{Polycheria.obtusa}}{a numeric vector giving individual counts for this species}
    \item{\code{Amphilochus.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Gitanopsis.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Ampithoe}}{a numeric vector giving individual counts for this species}
    \item{\code{Ceina.sp}}{a numeric vector giving individual counts for this species}
    \item{\code{Colomastix.subcastellata}}{a numeric vector giving individual counts for this species}
    \item{\code{Corophiidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Aora.Aoroides}}{a numeric vector giving individual counts for this species}
    \item{\code{Corophium.acutum}}{a numeric vector giving individual counts for this species}
    \item{\code{Gammaropsis.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Gammaropsis.tawahi}}{a numeric vector giving individual counts for this species}
    \item{\code{Paracorophium.excavatum}}{a numeric vector giving individual counts for this species}
    \item{\code{Ericthonius.brasiliensis}}{a numeric vector giving individual counts for this species}
    \item{\code{Ericthonius.sp}}{a numeric vector giving individual counts for this species}
    \item{\code{Microdeutopus.apopo}}{a numeric vector giving individual counts for this species}
    \item{\code{Neocyproidea.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Atylus.sp}}{a numeric vector giving individual counts for this species}
    \item{\code{Atylus.taupo}}{a numeric vector giving individual counts for this species}
    \item{\code{Paradexamine}}{a numeric vector giving individual counts for this species}
    \item{\code{Bircenna.fulva}}{a numeric vector giving individual counts for this species}
    \item{\code{Oradarea.novaezelandiae}}{a numeric vector giving individual counts for this species}
    \item{\code{Eusiridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Maera.incerta}}{a numeric vector giving individual counts for this species}
    \item{\code{Metaceradocus.whakatane}}{a numeric vector giving individual counts for this species}
    \item{\code{Melita.festiva}}{a numeric vector giving individual counts for this species}
    \item{\code{Parapherusa.crassipes}}{a numeric vector giving individual counts for this species}
    \item{\code{Hyale.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Panoploea.spinosa}}{a numeric vector giving individual counts for this species}
    \item{\code{Ischyrocerus}}{a numeric vector giving individual counts for this species}
    \item{\code{Ventojassa}}{a numeric vector giving individual counts for this species}
    \item{\code{Leucothoe.trailli}}{a numeric vector giving individual counts for this species}
    \item{\code{Liljeborgia.akaroica}}{a numeric vector giving individual counts for this species}
    \item{\code{Lys..sp..1..Ambasia.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Lys..sp..2}}{a numeric vector giving individual counts for this species}
    \item{\code{Stomacontion.hurleyi}}{a numeric vector giving individual counts for this species}
    \item{\code{Nihotunga.ilika}}{a numeric vector giving individual counts for this species}
    \item{\code{Iphinotus.typicus}}{a numeric vector giving individual counts for this species}
    \item{\code{Brolginae}}{a numeric vector giving individual counts for this species}
    \item{\code{Podocerus.karu}}{a numeric vector giving individual counts for this species}
    \item{\code{Seba.typica}}{a numeric vector giving individual counts for this species}
    \item{\code{Tetradeion.crassum.}}{a numeric vector giving individual counts for this species}
    \item{\code{Stenothoe.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Alpheus.socialis}}{a numeric vector giving individual counts for this species}
    \item{\code{Pagurus.sp...Hermit.crabs.}}{a numeric vector giving individual counts for this species}
    \item{\code{Petrolisthes.novaezelandiae}}{a numeric vector giving individual counts for this species}
    \item{\code{Halicarcinus.ovatus}}{a numeric vector giving individual counts for this species}
    \item{\code{Notomithrax.minor}}{a numeric vector giving individual counts for this species}
    \item{\code{Leptomithrax.garricki}}{a numeric vector giving individual counts for this species}
    \item{\code{Pilumnus.novaezelandiae}}{a numeric vector giving individual counts for this species}
    \item{\code{Halacaridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Pycnogonida}}{a numeric vector giving individual counts for this species}
    \item{\code{Acrocirridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Ampharetiidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Capitellidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Chaetopteridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Chrysopatellidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Cirratulidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Eunicidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Dorvillea}}{a numeric vector giving individual counts for this species}
    \item{\code{Eunice}}{a numeric vector giving individual counts for this species}
    \item{\code{Lysidice}}{a numeric vector giving individual counts for this species}
    \item{\code{Pherusa.parmata}}{a numeric vector giving individual counts for this species}
    \item{\code{Glyceridae}}{a numeric vector giving individual counts for this species}
    \item{\code{Lumbrinereidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Maldanidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Nereididae}}{a numeric vector giving individual counts for this species}
    \item{\code{Protoaricia.}}{a numeric vector giving individual counts for this species}
    \item{\code{Armandia}}{a numeric vector giving individual counts for this species}
    \item{\code{Phyllodocidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Polynoidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Sabellariidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Sabellidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Idyanthyrsus.sp.}}{a numeric vector giving individual counts for this species}
    \item{\code{Scalibregmidae..Hyboscolex.}}{a numeric vector giving individual counts for this species}
    \item{\code{Serpulidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Spirorbinae}}{a numeric vector giving individual counts for this species}
    \item{\code{Spionidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Spionidae...Boccardia}}{a numeric vector giving individual counts for this species}
    \item{\code{Syllinae}}{a numeric vector giving individual counts for this species}
    \item{\code{Trypanosyllis}}{a numeric vector giving individual counts for this species}
    \item{\code{Eusyllinae}}{a numeric vector giving individual counts for this species}
    \item{\code{Exogoninae}}{a numeric vector giving individual counts for this species}
    \item{\code{Terebellidae}}{a numeric vector giving individual counts for this species}
    \item{\code{Nemerteans}}{a numeric vector giving individual counts for this species}
    \item{\code{Nematodes}}{a numeric vector giving individual counts for this species}
    \item{\code{Flatworm...Robust..entire.edge}}{a numeric vector giving individual counts for this species}
    \item{\code{Platyhelminthes.4}}{a numeric vector giving individual counts for this species}
    \item{\code{Sipunculid}}{a numeric vector giving individual counts for this species}
  }
}
\details{
Extract rows for a species table of a particular holdfast
}
\source{
  Data supplied by Justine Saunders
}
\examples{
data(saunders)

}
\keyword{datasets}
