\name{spitale}
\alias{spitale}
\docType{data}
\title{Counts of diatom species in springs of the Adamello-Brenta Nature Park}
\description{
  A dataset due to Spitale and Cantonati comprising abundances of
  different species of diatoms
}
\usage{data(spitale)}
\format{
  A count object
}
\source{
  Data kindly provided by Daniel Spitale
}
\references{
  D. Spitale and M. Cantonati 2011.  \dQuote{Understanding the natural
    variability of diatom assemblages in springs of the Adamello-Brenta
    Nature Park (south-eastern Alps) on a temporal scale}.
  \emph{Fundamental Applied Limnology} volume 179/2, pp137--149
}
\examples{
data(spitale)
summary(spitale)
}   
\keyword{datasets}
