\name{logS1}
\alias{logS1}
\docType{data}
\title{logarithms of Stirling numbers of the first kind}
\description{
Natural logarithms of Stirling numbers of the first kind, used by
function \code{logkda()}.
}
\usage{logS1}
\format{
A 100-by-100 matrix of logs of Stirling numbers of the first kind
}
\source{
Calculated by Maple
}
\seealso{\code{\link{etienne}}}
\examples{
exp(logS1[1:5,1:5])
}
\keyword{datasets}
