\name{sahfos}
\alias{sahfos}
\docType{data}
\title{Biodiversity dataset provided by SAHFOS}
\description{
Species counts in the North Atlantic
}
\usage{data(sahfos)}
\source{
 \url{http://192.171.163.165/data.htm}
}
\references{
  Warner AJ and Hays GC 1994.  \dQuote{Sampling by the Continuous Plankton Recorder Survey}. Progress in Oceanography, 34: 237-256
}
\examples{
data(sahfos)
preston(sahfos)
}
\keyword{datasets}
